Refactor osmcha-stats
